requirements.txt
MetaTrader5>=5.0.0
pandas>=2.0
numpy>=1.25
python-dotenv
ta
pyyaml
uvicorn
fastapi
loguru
requests
